+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 * `--sources` to show definitions' source dictionary
+* `--from` to only show results from a given source dictionary
 
 ### Changed